endpoints to update quantity and delete user ingredients, add more tests

+++ src/main/java/com/kitchenapp/kitchenappapi/dto/NewMeasurementDTO.java
@@ -13,5 +13,5 @@ import lombok.Setter;
 public class NewMeasurementDTO {
     private String name;
     private String metricUnit;
-    private int metricQuantity;
+    private double metricQuantity;
 }
